+++ CHANGES.rst
@@ -1,3 +1,32 @@
+0.6.1
+=====
+February 21st, 2019
+
+* Import platform in setup.py
+
+0.6.0
+=====
+February 21st, 2019
+
+* Add support for Windows via winkerberos (#30)
+
+Thanks to @ryan-pip for this release
+
+0.5.1
+=====
+May 18th, 2018
+
+* Fixes for DIGEST-MD5, server auth has been completed and is now tested
+* Both DIGEST-MD5 and CRAM-MD5 properly set the complete flag now
+
+Thanks to @ceache and @bjmb for this release
+
+0.5.0
+=====
+March 8th, 2018
+
+* Added EXTERNAL mechanism
+
 0.4.1
 =====
 March 8th, 2018